refs(metric_alerts): Use `date_started` and `date_detected` correctly.
Previously we've treated `date_started` and `date_detected` as the same value, but we should
differentiate between these. `date_started` is the start of the bucket that created an alert.
`date_detected` is the end of that bucket, and is the date that the trigger conditions were met for
the alert. We also have `date_added`, which is the time we actually created it.

This corrects the subscription processor to pass these dates correctly. We also rename the `STARTED`
activity to `DETECTED`, and the old `DETECTED` activity to `CREATED`. This should have no impact on
the FE since we don't return those specific names, but more accurately represents the purpose of
those activities.

Relies on an upcoming frontend change to use `date_started` to render the started dot on the graph.